requirements.txt
backports.ssl-match-hostname==3.4.0.2
Django==1.8.3
djangorestframework==3.2.0
docker-py==1.3.1
requests==2.7.0
websocket-client==0.32.0
pytz==2015.4
git+git://github.com/akalipetis/djoser.git@hotfix/setup-djoser-urls
